Warning fixes:
not complete, but I stopped pulling the thread before the whole
project unravelled 8)

(1) team and player numbers should be unsigned.

(2) sizes of gui elements should usually be unsigned.

This probably introduced some bugs.

+++ src/unit.cpp
@@ -652,7 +652,7 @@ bool unit::matches_filter(const config& cfg) const
 		}
 	}
 
-	if(side.empty() == false && this->side() != atoi(side.c_str()))
+	if(side.empty() == false && this->side() != (unsigned)atoi(side.c_str()))
 	  {
 		if(std::find(side.begin(),side.end(),',') != side.end()) {
 			const std::vector<std::string>& vals = utils::split(side);
